+++ src/peer_manager/mod.rs
@@ -886,6 +886,8 @@ impl PeerManager {
         transport.max_prefixes = config.max_prefixes;
         transport.max_prefixes_ipv4 = config.max_prefixes_ipv4;
         transport.max_prefixes_ipv6 = config.max_prefixes_ipv6;
+        transport.max_prefixes_received_ipv4 = config.max_prefixes_received_ipv4;
+        transport.max_prefixes_received_ipv6 = config.max_prefixes_received_ipv6;
         transport.peer_group.clone_from(&config.peer_group);
         transport.md5_password.clone_from(&config.md5_password);
         transport.tcp_ao.clone_from(&config.tcp_ao);